Comments
Radera och flagga samt återställa

+++ index.php
@@ -321,8 +321,9 @@ if (key_exists('url',$_GET))
             }
             break;
         case "comment/delete":
-            if (key_exists('id',$_POST))
+            if (key_exists('id',$_POST) && key_exists('ReviewId',$_POST))
             {
+                echo "funkar?";
                 require_once "controller/Comments.Controller.php";
                 $controller = new CommentsController();
                 $controller->DeleteComment($_POST['id']);
@@ -337,7 +338,7 @@ if (key_exists('url',$_GET))
             }
             break;
         case "reply/delete":
-            if (key_exists('id',$_POST))
+            if (key_exists('id',$_POST) && key_exists('ReviewId',$_POST))
             {
                 require_once "controller/Comments.Controller.php";
                 $controller = new CommentsController();
@@ -382,6 +383,9 @@ if (key_exists('url',$_GET))
             $controller->ShowContactPage();
             break;
         default:
+            require_once "controller/Home.Controller.php";
+            $controller = new HomeController();
+            $controller->ShowHomePage();
             break;
     }
     exit();
@@ -532,6 +536,7 @@ function BooksRoute($action)
         default:
             break;
     }
+    exit();
 }
 
 function UserRoute($action)
@@ -555,6 +560,7 @@ function UserRoute($action)
             $controller->Logout(); //Loggar ut och förstör session
             break;
     }
+    exit();
 }
 function AdminRoute($action)
 {
@@ -595,6 +601,7 @@ function AdminRoute($action)
         default:
             break;
     }
+    exit();
 }
 
 function ModeratorRoute($action)
@@ -609,5 +616,6 @@ function ModeratorRoute($action)
         default:
             break;
     }
+    exit();
 }
 ?>
